Adjust how we handle absolute/relative paths and file contents (#434)
* Clarify absolute vs. relative paths and when we have and don't have file contents

In libraries.io, we get only relative paths with no file contents from GitHub
APIs. Relative path = relative to the project directory.

When scanning a local directory, we need to load file contents (and that
means we need to use paths that are open-able, so we need the full
path or at least relative to pwd, not project directory).

This PR is attempting to clarify this by always having "contents"
available in the local file case. To lazy-load the contents,
we pass around a FileInfo object instead of trying to keep
track of what a given filename-as-string is relative to and passing
around separate "contents" and "folder_path" parameters.