Avoid root method on result of 0x.parse_file_contents
Use locate() instead. Not sure why/when root is missing,
but apparently it can be in production.

Also, catch any exceptions in ivy_report? matcher,
since we aren't going to handle them at that point.
This does mean we may have worse error messages
when a file doesn't have ivy-report at the root,
since we won't even get around to trying to parse it.

Future solution might be to have a mode where
all files are expected to be parseable (and we might
not even do the match first).