Add ignored_files and fixes to ignored_dirs
 * there's now an ignored_files configuration option
 * ignored_dirs now matches entire relative path not only basename
 * test for ignored_dirs actually works
 * in identify_manifests, ignoring "foo" no longer ignores "foobar",
   but still ignores "foo/bar"

ignored_dirs previously matched the dir anywhere that the basename
occurred in the hierarchy, and had to be a single filename not a path.
This was in the "we have a local checkout" case.
But when identifying manifests without a local checkout, it
required the dir to be at the start of the path.

It now consistently works the same way, you _can_ ignore `spec/fixtures`
but ignoring only `fixtures` won't match that. This isn't compatible
with one of the cases, but seems more expected overall?

ignored_dirs test had a test that always succeeded even if ignored_dirs
didn't work, fixed that.

Added an ignored_files feature to allow ignoring certain files.